feature(svg-generator): Add config-dir option to svg-generator CLI
svg-generator CLI now takes an optional argument to specify where the svgGenerator
config object should be looked for, defaulting to previous behavior of looking
in the current working directory.

This allows to support use cases where separate instances of svg-generator shoud be
used in subdirectories of the target project (typically in a Nx monorepo setting).

Close #64